+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggplot2 (development version)
 
-* The helper function `is.waive()` is now exported to help extensions to work
+* The helper function `is.waiver()` is now exported to help extensions to work
   with `waiver()` objects (@arcresu, #6173).
 * (internal) Using `after_scale()` in the `Geom*$default_aes()` field is now
   evaluated in the context of data (@teunbrand, #6135)